I commit at 7:03 Add register a car function

+++ App/src/main/java/khoindn/swp391/be/app/model/Response/UsersResponse.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class UsersResponse {
-    long id;
+    int id;
     String hovaten;
     String email;
     String cccd;
